+++ recycler-extension/RELEASE_NOTES.md
@@ -4,7 +4,7 @@
 ##### Recycler-extension
 * ANDDEP-687 Changed "com.google.android.material:material" dependency
   from "api" to "implementation" type
-* ANDDEP-1038 Add `AnimatableViewHolder` and `BaseItemAnimator`
+* ANDDEP-1038 Add `AnimatableViewHolder` and `ViewHolderItemAnimator`
 ## 0.4.0
 ##### Recycler-extension
 * ANDDEP-266 DividerItemDecorator Update